[l10n] Add localization macros and refactor
