+++ bve/src/lib.rs
@@ -47,11 +47,13 @@
 
 pub use datatypes::*;
 
+#[macro_use]
+pub mod l10n;
+
 pub mod concurrency;
 pub mod data;
 mod datatypes;
 pub mod filesystem;
 mod iter;
-pub mod l10n;
 pub mod log;
 pub mod parse;
